Directory setup and basic GUI
A basic GUI has been created and files have been added that will hold each part of the app. Right now the different features include my lat and long, the iss lat and long, sunrise and sunset tracking, and the ability to send an email when the iss is potentially visible.