Updated bst.py
Created class binarytree with init, add, treeify, and search methods. init checks for a json file, if one does not exist, one is created. add creates a new data node for the json file then calls treeify. treeify sorts the new node into the correct location of the tree. search locates a node in the tree and returns the wanted data or None.